COM-995 Move docx generation to its own helper
