+++ src/helpers/file.js
@@ -1,29 +1,8 @@
 const fs = require('fs');
 const os = require('os');
 const path = require('path');
-const JSZip = require('jszip');
-const DocxTemplater = require('docxtemplater');
-const drive = require('../models/Google/Drive');
 
 const fsPromises = fs.promises;
-
-exports.generateDocx = async (params) => {
-  params.file.tmpFilePath = path.join(os.tmpdir(), 'template.docx');
-  await drive.downloadFileById(params.file);
-
-  const file = await fsPromises.readFile(params.file.tmpFilePath, 'binary');
-  const zip = new JSZip(file);
-  const doc = new DocxTemplater();
-  doc.loadZip(zip);
-  doc.setData(params.data);
-  doc.render();
-  const filledZip = doc.getZip().generate({ type: 'nodebuffer' });
-  const date = new Date();
-  const tmpOutputPath = path.join(os.tmpdir(), `template-filled-${date.getTime()}.docx`);
-  await fsPromises.writeFile(tmpOutputPath, filledZip);
-
-  return tmpOutputPath;
-};
 
 exports.createAndReadFile = async (stream, outputPath) => new Promise((resolve, reject) => {
   const tmpFile = fs.createWriteStream(outputPath);
